+++ rcs-protocol-vda5050/src/test/java/io/github/cooltheworld/rcs/protocol/vda5050/v3/architecture/PublicApiArchitectureTest.java
@@ -86,6 +86,8 @@ class PublicApiArchitectureTest {
             "Envelope3d",
             "Envelope3dData",
             "LocalizationType",
+            "LoadDimensions",
+            "LoadSpecification",
             "MaximumArrayLengths",
             "MaximumStringLengths",
             "MobileRobotClass",